ChatGPT quickly made a zarr-based 0.04x0.04 global map script. Haven't tested it yet.
