+++ src/LULUCF/scripts/vegetation_model/OLD_create_0_04deg_global_display_maps.py
