+++ src/LULUCF/scripts/vegetation_model/OLD_create_chunks_0_04x0_04deg.py
