+++ src/LULUCF/scripts/vegetation_model/OLD_create_global_0_04x0_04deg.py
